+++ src/components/ProfileDetails.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 // import ProfileContent from "./common/ProfileList";
-import ProfileImage from "../components/common/Profile/ProfileImage";
+import ProfileImage from "./common/Profile/ProfileImage";
 import AboutUs from "./AboutUs";
 
 const ProfileDetails = () => {
