--- src/data/educationDetails.ts
@@ -1,33 +1,0 @@
-interface EducationDetails {
-  degree: string;
-  fieldOfStudy: string;
-  institution: string;
-  yearOfGraduation: number;
-}
-
-export const educationDetails: EducationDetails[] = [
-  {
-    degree: "M Tech",
-    fieldOfStudy: "Computer Science and Engineering",
-    institution: "PBR Vits",
-    yearOfGraduation: 2023,
-  },
-  {
-    degree: "B Tech",
-    fieldOfStudy: "Computer Science and Engineering",
-    institution: "PBR Vits",
-    yearOfGraduation: 2019,
-  },
-  {
-    degree: "Intermediate",
-    fieldOfStudy: "M.P.C",
-    institution: "Srinivasa Junior College",
-    yearOfGraduation: 2015,
-  },
-  {
-    degree: "10th Grade",
-    fieldOfStudy: "General Studies",
-    institution: "Ratnam High School",
-    yearOfGraduation: 2013,
-  },
-];
